Rearranged common folder to move *_response.py and *_json.py to their respective new folders to make them easier to find.
Amended code to point to these scripts in their new folders.
